tests/wpt/css-tests/cssom-view-1_dev/xhtml1/ttwf-scrollintoview.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSSOM View - 6 - scrollIntoView tests</title>
    <meta charset="utf-8" />
    <link href="mailto:pwx.frontend@gmail.com" rel="author" title="Chris Wu" />
	<link href="http://www.w3.org/TR/cssom-view/#dom-element-scrollintoview" rel="help" />
    <meta content="dom" name="flags" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style type="text/css">
    	#phitem{height: 2000px;}
    	#viewitem{width:300px;height:200px;margin:0 auto;background-color: lightgreen;}
    </style>
</head>
<body>
    <div id="myDiv">
    	<div id="phitem"></div>
    	<div id="viewitem"></div>
    </div>
    <div id="log"></div>
    <script>
		var asytest = async_test('check scrollIntoView');
		asytest.step(function(){
			var viewitem = document.getElementById('viewitem');
			viewitem.scrollIntoView();
			assert_true(window.pageYOffset + window.innerHeight &gt; viewitem.offsetTop, "The element isn't in the viewport")
		})
		asytest.done()
    </script>


</body></html>
